refactor: InputView, OutputView 변경 적용시키기
